updated createLevelSelect, tweaked animationctrl because of dereference error, fixed method signatures of uidata

+++ source/UIData.hpp
@@ -23,7 +23,7 @@ public:
 	float width;
 	float height;
 
-	bool init(std::string uiKey, std::string bg, UIDataType t, int x, int y, int w, int h)
+	bool init(std::string uiKey, std::string bg, UIDataType t, float x, float y, float w, float h)
 	{
 		this->UIKey = uiKey;
 		this->uiBackgroundKey = bg;
@@ -37,7 +37,7 @@ public:
 
 	UIData() : Data() {}
 
-	static std::shared_ptr<UIData> alloc(std::string uiKey, std::string bg, UIDataType t, int x, int y, int w, int h) {
+	static std::shared_ptr<UIData> alloc(std::string uiKey, std::string bg, UIDataType t, float x, float y, float w, float h) {
 		std::shared_ptr<UIData> result = std::make_shared<UIData>();
 		return (result->init(uiKey, bg, t, x, y, w, h) ? result : nullptr);
 	}
@@ -64,7 +64,7 @@ public:
 
 	virtual std::shared_ptr<cugl::Node> dataToNode(std::shared_ptr<GenericAssetManager> assets) override;
 
-	bool init(std::string uiKey, std::string bg, int x, int y, int w, int h, std::shared_ptr<ButtonAction> ba, std::string bLabel)
+	bool init(std::string uiKey, std::string bg, float x, float y, float w, float h, std::shared_ptr<ButtonAction> ba, std::string bLabel)
 	{
 		UIData::init(uiKey, bg, UIDataType::BUTTON, x, y, w, h);
 		buttonAction = ba;
@@ -72,7 +72,7 @@ public:
 		return true;
 	}
 
-	static std::shared_ptr<ButtonUIData> alloc(std::string uiKey, std::string bg, int x, int y, int w, int h, std::shared_ptr<ButtonAction> ba, std::string bLabel) {
+	static std::shared_ptr<ButtonUIData> alloc(std::string uiKey, std::string bg, float x, float y, float w, float h, std::shared_ptr<ButtonAction> ba, std::string bLabel) {
 		std::shared_ptr<ButtonUIData> result = std::make_shared<ButtonUIData>();
 		return (result->init(uiKey, bg, x, y, w, h, ba, bLabel) ? result : nullptr);
 	}
@@ -89,11 +89,11 @@ public:
 
 	virtual std::shared_ptr<cugl::Node> dataToNode(std::shared_ptr<GenericAssetManager> assets) override;
 
-	static std::shared_ptr<TextUIData> alloc(std::string uiKey, std::string bg, int x, int y, int w, int h, std::string tv, std::string fKey) {
+	static std::shared_ptr<TextUIData> alloc(std::string uiKey, std::string bg, float x, float y, float w, float h, std::string tv, std::string fKey) {
 		std::shared_ptr<TextUIData> result = std::make_shared<TextUIData>();
 		return (result->init(uiKey, bg, x, y, w, h, tv, fKey) ? result : nullptr);
 	}
-	bool init(std::string uiKey, std::string bg, int x, int y, int w, int h, std::string tv, std::string fKey)
+	bool init(std::string uiKey, std::string bg, float x, float y, float w, float h, std::string tv, std::string fKey)
 	{
 		UIData::init(uiKey, bg, UIDataType::TEXT, x, y, w, h);
 		textValue = tv;
@@ -112,11 +112,11 @@ public:
 
 	virtual std::shared_ptr<cugl::Node> dataToNode(std::shared_ptr<GenericAssetManager> assets) override;
 
-	static std::shared_ptr<ImageUIData> alloc(std::string uiKey, std::string bg, int x, int y, int w, int h, std::string tKey) {
+	static std::shared_ptr<ImageUIData> alloc(std::string uiKey, std::string bg, float x, float y, float w, float h, std::string tKey) {
 		std::shared_ptr<ImageUIData> result = std::make_shared<ImageUIData>();
 		return (result->init(uiKey, bg, x, y, w, h, tKey) ? result : nullptr);
 	}
-	bool init(std::string uiKey, std::string bg, int x, int y, int w, int h, std::string tKey)
+	bool init(std::string uiKey, std::string bg, float x, float y, float w, float h, std::string tKey)
 	{
 		UIData::init(uiKey, bg, UIDataType::IMAGE, x, y, w, h);
 		textureKey = tKey;
